Explicitly unwrap dial backoff errors to tag metrics
The error kind returned as a result of dial backoff is wrapped in such a
way that the regular go SDK `Unwrap` and `Is` will not surface it.

Considering it is a popular error, and occurs when the peered node addrs
are no longer dialable, write a bespoke unwrapping for it in order to
tag metrics with the right error kind.