Implement automatic peer address discovery via IPFS DHT
Add the capability to automatically discover and refresh addresses for a
configured peer ID when the configured peer multiaddr only contains the
peer ID. In such case, the address associated to the peer ID are
automatically discovered by performing peer routing lookup on the IPFS
DHT in client mode. The address for such peers are then kept up-to-date
by periodic lookups.

This feature allows `cassette` to cope with peers that keep the same ID
but change their IP address frequently.