Delete recipients on Circuit breaker opening
When a peered node changes address or is no longer accessible it remains
in the list of recipients indefinitely with circuit breaker ping ponging
between different states until the pod is restarted.

The changes here remove a recipient if it misbehaves such that it opens
a the circuit breaker. The sensitivity to misbehaviour is then
configurable by setting circuit breaker open/half-open characteristics.

The recipients of broadcast are expected to peer with Cassette. Peering
attempts to re-connect periodically. So as long as recipients have
peered with Cassete they should get re-added as normal.